Merge branches.

+++ src/main/java/bcamultiblocks/tile/TileMultiBlockFake.java
@@ -4,41 +4,41 @@ import net.minecraft.nbt.NBTTagCompound;
 
 public class TileMultiBlockFake extends TileMultiBlock {
 
-	private String originalName;
-	private int originalMeta;
+    private String originalName;
+    private byte originalMeta;
 
-	@Override
-	public void writeToNBT(NBTTagCompound compound) {
-		super.writeToNBT(compound);
-		compound.setString("BlockName", originalName);
-		compound.setByte("BlockMeta", (byte) originalMeta);
-	}
+    @Override
+    public void writeToNBT(NBTTagCompound compound) {
+        super.writeToNBT(compound);
+        compound.setString("BlockName", originalName);
+        compound.setByte("BlockMeta", originalMeta);
+    }
 
-	@Override
-	public void readFromNBT(NBTTagCompound compound) {
-		super.readFromNBT(compound);
-		originalName = compound.getString("BlockName");
-		originalMeta = compound.getByte("BlockMeta");
-	}
+    @Override
+    public void readFromNBT(NBTTagCompound compound) {
+        super.readFromNBT(compound);
+        originalName = compound.getString("BlockName");
+        originalMeta = compound.getByte("BlockMeta");
+    }
 
-	public void setBlock(String name, int meta) {
-		setOriginalName(name);
-		setOriginalMeta(meta);
-	}
+    public void setBlock(String name, byte meta) {
+        setOriginalName(name);
+        setOriginalMeta(meta);
+    }
 
-	public String getOriginalName() {
-		return originalName;
-	}
+    public String getOriginalName() {
+        return originalName;
+    }
 
-	public void setOriginalName(String originalName) {
-		this.originalName = originalName;
-	}
+    public void setOriginalName(String originalName) {
+        this.originalName = originalName;
+    }
 
-	public int getOriginalMeta() {
-		return originalMeta;
-	}
+    public int getOriginalMeta() {
+        return originalMeta;
+    }
 
-	public void setOriginalMeta(int originalMeta) {
-		this.originalMeta = originalMeta;
-	}
+    public void setOriginalMeta(byte originalMeta) {
+        this.originalMeta = originalMeta;
+    }
 }
